requirements.txt
Flask
flask-cors
reportlabs